refactor(tool-git): implement single responsibility pattern
- Split monolithic main.go (258 lines) into focused modules:
  - main.go: Clean entry point with plugin metadata (60 lines)
  - git_tool.go: Execute method and command routing
  - git_config.go: Git configuration management with sensible defaults
  - git_aliases.go: Comprehensive Git aliases installation
  - git_status.go: Configuration status and reporting

- Enhanced functionality:
  - Improved configuration validation and error handling
  - Enhanced status reporting with detailed configuration info
  - Better user feedback with emojis and clear messages
  - Comprehensive alias management with 30+ useful shortcuts

- Follows same modular pattern as package-manager-apt
- Each file maintains single responsibility
- Improved maintainability and testability
